AutoGrader.TestUnit behaviour (auto_grader v0.1.0)

A behavior for defining test units.

The run/2 callback is called to run the test unit, and must return a {score, max_score} tuple.

The potential context generated from setup units will be passed as a second argument.

A coefficient/0 callback can be defined to modify the coefficient of the test unit (defaults to 1).

A bonus?/0 callback can be defined to treat the test unit as a bonus exercice (defaults to false).

Callbacks

bonus?()
(optional)

@callback bonus?() :: boolean()

Whether or not this test unit is a bonus one.

coefficient()
(optional)

@callback coefficient() :: number()

The coefficient to apply to this test unit.

run(submission_path, context)

@callback run(submission_path :: binary(), context :: map()) :: {number(), number()}

Runs the test unit.

Returns a {score, max_score} tuple.

Functions

bonus?()

coefficient()